Update travis config and test script

+++ tests/test-in-docker-image.sh
@@ -2,12 +2,13 @@
 
 SOURCE="${BASH_SOURCE[0]}"
 RDIR="$( dirname "$SOURCE" )"
-SUDO=`which sudo 2> /dev/null`
+SUDO=$(which sudo 2> /dev/null)
 SUDO_OPTION=""
 #SUDO_OPTION="--sudo"
 OS_TYPE=${1:-}
 OS_VERSION=${2:-}
 ANSIBLE_VERSION=${3:-}
+export ANSIBLE_VERSION
 
 ANSIBLE_VAR=""
 ANSIBLE_INVENTORY="tests/inventory"
@@ -17,24 +18,25 @@ ANSIBLE_LOG_LEVEL="-v"
 APACHE_CTL="apache2ctl"
 
 # if there wasn't sudo then ansible couldn't use it
-if [ "x$SUDO" == "x" ];then
+if [ -z "$SUDO" ];then
     SUDO_OPTION=""
 fi
 
 if [ "${OS_TYPE}" == "centos" ];then
     APACHE_CTL="apachectl"
+    export APACHE_CTL
     if [ "${OS_VERSION}" == "7" ];then
         ANSIBLE_VAR="apache_use_service=False"
     fi
 fi
 
-ANSIBLE_EXTRA_VARS=""
-if [ "${ANSIBLE_VAR}x" == "x" ];then
-    ANSIBLE_EXTRA_VARS=" -e \"${ANSIBLE_VAR}\" "
+ANSIBLE_EXTRA_VARS=()
+if [ -n "${ANSIBLE_VAR}" ];then
+    ANSIBLE_EXTRA_VARS=(-e "${ANSIBLE_VAR}")
 fi
 
 
-cd $RDIR/..
+cd "${RDIR}/.." || exit 1
 printf "[defaults]\nroles_path = ../:roles" > ansible.cfg
 printf "" > ssh.config
 
@@ -55,7 +57,10 @@ function install_ansible_devel() {
 
 echo "TEST: building ansible"
 
-yum -y install PyYAML python-paramiko python-jinja2 python-httplib2 rpm-build make python2-devel asciidoc patch wget 2>&1 >/dev/null || (echo "Could not install ansible yum dependencies" && exit 2 )
+if ! yum -y install PyYAML python-paramiko python-jinja2 python-httplib2 rpm-build make python2-devel asciidoc patch wget 2>&1 >/dev/null; then
+    echo "Could not install ansible yum dependencies"
+    exit 2
+fi
 rm -Rf ansible
 git clone https://github.com/ansible/ansible --recursive ||(echo "Could not clone ansible from Github" && exit 2 )
 cd ansible
@@ -104,15 +109,15 @@ function test_playbook_syntax(){
 }
 
 function test_playbook_check(){
-    echo "TEST: ansible-playbook -i ${ANSIBLE_INVENTORY} ${ANSIBLE_PLAYBOOk} ${ANSIBLE_LOG_LEVEL} --connection=local ${SUDO_OPTION} ${ANSIBLE_EXTRA_VARS} --check"
+    echo "TEST: ansible-playbook -i ${ANSIBLE_INVENTORY} ${ANSIBLE_PLAYBOOk} ${ANSIBLE_LOG_LEVEL} --connection=local ${SUDO_OPTION} ${ANSIBLE_EXTRA_VARS[*]} --check"
 
-    ansible-playbook -i ${ANSIBLE_INVENTORY} ${ANSIBLE_PLAYBOOk} ${ANSIBLE_LOG_LEVEL} --connection=local ${SUDO_OPTION} ${ANSIBLE_EXTRA_VARS} --check ||(echo "playbook check failed" && exit 2 )
+    ansible-playbook -i ${ANSIBLE_INVENTORY} ${ANSIBLE_PLAYBOOk} ${ANSIBLE_LOG_LEVEL} --connection=local ${SUDO_OPTION} "${ANSIBLE_EXTRA_VARS[@]}" --check ||(echo "playbook check failed" && exit 2 )
 
 }
 
 function test_playbook(){
-    echo "TEST: ansible-playbook -i ${ANSIBLE_INVENTORY} ${ANSIBLE_PLAYBOOk} ${ANSIBLE_LOG_LEVEL} --connection=local ${SUDO_OPTION} ${ANSIBLE_EXTRA_VARS}"
-    ansible-playbook -i ${ANSIBLE_INVENTORY} ${ANSIBLE_PLAYBOOk} ${ANSIBLE_LOG_LEVEL} --connection=local ${SUDO_OPTION} ${ANSIBLE_EXTRA_VARS} ||(echo "first ansible run failed" && exit 2 )
+    echo "TEST: ansible-playbook -i ${ANSIBLE_INVENTORY} ${ANSIBLE_PLAYBOOk} ${ANSIBLE_LOG_LEVEL} --connection=local ${SUDO_OPTION} ${ANSIBLE_EXTRA_VARS[*]}"
+    ansible-playbook -i ${ANSIBLE_INVENTORY} ${ANSIBLE_PLAYBOOk} ${ANSIBLE_LOG_LEVEL} --connection=local ${SUDO_OPTION} "${ANSIBLE_EXTRA_VARS[@]}" ||(echo "first ansible run failed" && exit 2 )
 
 #    echo "TEST: idempotence test! Same as previous but now grep for changed=0.*failed=0"
 #    ansible-playbook -i ${ANSIBLE_INVENTORY} ${ANSIBLE_PLAYBOOk} ${ANSIBLE_LOG_LEVEL} --connection=local ${SUDO_OPTION} ${ANSIBLE_EXTRA_VARS} || grep -q 'changed=0.*failed=0' && (echo 'Idempotence test: pass' ) || (echo 'Idempotence test: fail' && exit 1)
